fix(cluster/namespace): 集群ns删除时判断是否存在工具

+++ ui/src/app/business/cluster/cluster-detail/namespace/namespace-delete/namespace-delete.component.ts
@@ -2,9 +2,11 @@ import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {KubernetesService} from '../../../kubernetes.service';
 import {ActivatedRoute} from '@angular/router';
 import {CommonAlertService} from '../../../../../layout/common-alert/common-alert.service';
+import {ModalAlertService} from '../../../../../shared/common-component/modal-alert/modal-alert.service';
 import {TranslateService} from '@ngx-translate/core';
 import {Cluster} from '../../../cluster';
 import {AlertLevels} from '../../../../../layout/app-alert/alert';
+import {ToolsService} from '../../tools/tools.service';
 
 @Component({
     selector: 'app-namespace-delete',
@@ -21,7 +23,9 @@ export class NamespaceDeleteComponent implements OnInit {
 
     constructor(private service: KubernetesService, private route: ActivatedRoute,
                 private commonAlertService: CommonAlertService,
-                private translateService: TranslateService) {
+                private modalAlertService: ModalAlertService,
+                private translateService: TranslateService,
+                private toolsService: ToolsService) {
     }
 
     ngOnInit(): void {
@@ -34,13 +38,28 @@ export class NamespaceDeleteComponent implements OnInit {
     }
 
     onSubmit() {
-        this.service.deleteNamespace(this.currentCluster.name, this.namespace).subscribe(res => {
-            this.opened = false;
-            this.commonAlertService.showAlert(this.translateService.instant('APP_DELETE_SUCCESS'), AlertLevels.SUCCESS);
-            this.deleted.emit();
-        }, error => {
-            this.opened = false;
-            this.commonAlertService.showAlert(error.error.message, AlertLevels.ERROR);
+        let exitStr: string = '';
+        this.toolsService.list(this.currentCluster.name).subscribe(d => {
+            if (d) {
+                for (const tool of d) {
+                    if (tool.vars["namespace"] === this.namespace && tool.status !== 'Waiting') {
+                        exitStr += tool.name + ',';
+                    }
+                }
+                if (exitStr === '') {
+                    this.service.deleteNamespace(this.currentCluster.name, this.namespace).subscribe(res => {
+                        this.opened = false;
+                        this.commonAlertService.showAlert(this.translateService.instant('APP_DELETE_SUCCESS'), AlertLevels.SUCCESS);
+                        this.deleted.emit();
+                    }, error => {
+                        this.opened = false;
+                        this.commonAlertService.showAlert(error.error.message, AlertLevels.ERROR);
+                    });
+                } else {
+                    exitStr = exitStr.substring(0, exitStr.length - 1);
+                    this.modalAlertService.showAlert(this.translateService.instant('APP_DELETE_NS_ERR') + exitStr, AlertLevels.ERROR);
+                }
+            }
         });
     }
